Modularize the project with the data and domain layers.
